[refactor]: move expression len out of public API into core
